get git config
